+++ src/components/App.js
@@ -10,9 +10,9 @@ import Home from '../components/Home';
 import Login from '../components/Login';
 import Register from '../components/Register';
 import ProfileStepper from '../components/ProfileStepper';
-import Profile from '../components/ProfileMui';
+import Profile from '../components/Profile';
 import ProfileSettings from '../components/ProfileSettings';
-import FinanceMainView from '../components/Finance/FinanceMainViewMui';
+import FinanceMainView from '../components/Finance/FinanceMainView';
 import AssetBreakdown from '../components/Finance/AssetBreakdown';
 import LiabilityEditor from '../components/Finance/LiabilityEditor';
